Add files via upload
This is a peer graded assignment for IBM Machine Learning Professional Certificate Program in Coursera.
I am doing EDA of the dataset that I have, then I run Linear Regression, Ridge and Lasso. At the end it is seen that Ridge and Lasso are performing slightly better than Linear Regression, although the differences are not really wide.